Moved subcomponents to OrderDetailWidget
