+++ src/components/@widgets/OrderDetailWidget/hooks/useFormattedTokenAmount.ts
