+++ src/components/@widgets/OrderDetailWidget/subcomponents/FilledAndStatus/FilledAndStatus.styles.tsx
