+++ src/components/@widgets/OrderDetailWidget/subcomponents/InfoSection/InfoSection.styles.tsx
