+++ src/components/@widgets/OrderDetailWidget/subcomponents/OrderStatusInfo/OrderStatusInfo.styles.tsx
