+++ src/components/@widgets/OrderDetailWidget/subcomponents/RecipientAndStatus/RecipientAndStatus.styles.tsx
